--- src/bin/xkpasswd.rs
@@ -1,3 +1,0 @@
-fn main() {
-    println!("{}", xkpasswd::gen_passwd(3));
-}
