+++ src/main.rs
@@ -1,0 +1,5 @@
+use xkpasswd::prelude::*;
+
+fn main() {
+    println!("{}", gen_passwd(3));
+}
