--- src/utils.rs
@@ -1,4 +1,0 @@
-pub fn set_panic_hook() {
-    #[cfg(feature = "console_error_panic_hook")]
-    console_error_panic_hook::set_once();
-}
